requirements.txt
streamlit==1.32.0
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
watchdog==3.0.0
supervisor==4.2.5
logrotate==3.21.0